2024/6/11 잔버그 수정 및 sql 수정문 완료
admin 영역과 Readme의 추가할예정

+++ src/main/java/com/JYProject/project/repository/BoardRepository/BoardMapperRepository.java
@@ -11,7 +11,7 @@ import java.util.Map;
 @Mapper
 public interface BoardMapperRepository {
 
-     int insertBoard(Board board);
+     Long insertBoard(Board board);
      Board selectBoardDetail(Long boardId);
      int updateBoard(Board board);
      int deleteBoard(Long boardId);
